Moved the keyboard functions into a separate file
