--- src/timers.lua
@@ -1,5 +1,0 @@
-return {
-    left = 0,
-    right = 0,
-    down = 0
-}
